apps/protocol/requirements.txt
typer>=0.9
pydantic>=2.0
pyyaml>=6.0
